Add a more explicit error when no `parts` returned (#180)
* Add a more explicit error when no `parts` returned

When a response has been blocked, the API will return a `Candidate`
(with `finish_reason`, `safety_ratings`, etc) that does not have
`Part`s.  Given blocking is common enough, and we can identify this
case, I've added a more explicit (and helpful?) error message to guide
users if this happens.

* Run latest version of `black`

These appear to be new changes, unrelated to this PR, but sure why not.